+++ backend/interviewer-service/src/main/java/com/hackerrank/interviewer_service/DTO/ScheduleInterviewDTO.java
@@ -6,6 +6,6 @@ import java.sql.Date;
 
 @Data
 public class ScheduleInterviewDTO {
-    private Integer enrollId;
+    private Long enrollId;
     private Date interviewDate;
 }
