+++ backend/interviewer-service/src/main/java/com/hackerrank/interviewer_service/DTO/UpdateTestScoreDTO.java
@@ -3,12 +3,11 @@ package com.hackerrank.interviewer_service.DTO;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateTestScoreDTO {
-    private Integer jobId;
-    private Integer candidateId;
+    private Long jobId;
+    private Long candidateId;
     private String TestScore;
 }
