+++ backend/interviewer-service/src/main/java/com/hackerrank/interviewer_service/Repository/InterviewRecordRepository.java
@@ -4,6 +4,6 @@ import com.hackerrank.interviewer_service.Entity.Enrollment;
 import com.hackerrank.interviewer_service.Entity.InterviewRecord;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InterviewRecordRepository extends JpaRepository<InterviewRecord,Integer> {
+public interface InterviewRecordRepository extends JpaRepository<InterviewRecord, Long> {
     InterviewRecord findByEnrollment(Enrollment e);
 }
